# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Black Pear Joggers API",
        "contact": {
            "name": "Paul Evans",
            "url": "https://bpj.org.uk",
            "email": "paul3vanz@gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/membership/{firstName}/{lastName}/{dateOfBirth}": {
            "get": {
                "tags": [
                    "Membership"
                ],
                "summary": "Check UKA membership by name and date of birth",
                "operationId": "e2ef527ec12dec194c23c74d16726951",
                "parameters": [
                    {
                        "name": "firstName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dateOfBirth",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/membership/{urn}": {
            "get": {
                "tags": [
                    "Membership"
                ],
                "summary": "Check UKA membership by URN",
                "operationId": "74e10761e72d7828fef467ea4d81a74e",
                "parameters": [
                    {
                        "name": "urn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clubs": {
            "get": {
                "tags": [
                    "Membership"
                ],
                "summary": "Get all UKA clubs",
                "operationId": "c5bafd06fea8ce76b45d200e862914b8",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clubs/{clubId}/members": {
            "get": {
                "tags": [
                    "Membership"
                ],
                "summary": "Get all UKA members of a club",
                "operationId": "0106431b63d3857aaaf20a6b95849286",
                "parameters": [
                    {
                        "name": "clubId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/parkrun/alphabet": {
            "get": {
                "tags": [
                    "Parkrun"
                ],
                "summary": "Get parkrun alphabet/challenge progress for a year",
                "operationId": "1307fa68860ead9a8bd28e91dcae3a49",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "challenge",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}